Add OnceInitScratch for concurrency-safe initialization
This adds OnceInitScractch for safe concurrent initialization of
scratchspaces. A common use for scratchspaces is to set up a
directory, fill it with some files, and then hand it off to
some external process. If the initialization is not made
concurrency-safe, this can cause problems (either by providing
the directory to the external process to early or by trying
an initialization twice that may not be idempotent). This
provides an opt-in concurrency safe interface that pidlocks
the initial creation of the scratch-space until after a
provided initializer has run. It is superficially similar
to https://github.com/JuliaLang/julia/pull/59002, but keyed
on scratchspaces and does not try to ensure that the initializer
is the same for all instances - doing this correctly is up to
the user. It simply provides mutual exclusion for the running
initializers.